+++ sdk/core/azure-core/inc/azure/core/http/curl_transport.hpp
@@ -40,6 +40,13 @@ namespace Azure { namespace Core { namespace Http {
   struct CurlTransportOptions final
   {
     /**
+     * @brief Default Maximum time in seconds that you allow the connection phase to the server to
+     * take.
+     *
+     */
+    AZ_CORE_DLLEXPORT static const long DefaultConnectionTimeout = 300;
+
+    /**
      * @brief The string for the proxy is passed directly to the libcurl handle without any parsing
      *
      * @remark No validation for the string is done by the Azure SDK. More about this option:
@@ -88,6 +95,25 @@ namespace Azure { namespace Core { namespace Http {
      *
      */
     CurlTransportSslOptions SslOptions;
+
+    /**
+     * @brief When true, libcurl will not use any functions that install signal handlers or any
+     * functions that cause signals to be sent to the process.
+     *
+     * @details This option is here to allow multi-threaded unix applications to still set/use all
+     * timeout options etc, without risking getting signals.
+     *
+     */
+    bool NoSignal = false;
+
+    /**
+     * @brief Contain the maximum time in seconds that you allow the connection phase to the server
+     * to take.
+     *
+     * @details This only limits the connection phase, it has no impact once it has connected.
+     *
+     */
+    long ConnectionTimeout = DefaultConnectionTimeout;
   };
 
   /**